src/library-layer/pyproject.toml
[tool.poetry]
name = "library-layer"
version = "0.1.0"
description = "Shared Lambda layer: runtime deps + steampulse framework"
authors = ["SteamPulse <hello@steampulse.io>"]
packages = [{ include = "library_layer" }]

[tool.poetry.dependencies]
python = "^3.12"
httpx = ">=0.27.0"
psycopg2-binary = ">=2.9.9"
boto3 = ">=1.34.0"
anthropic = ">=0.40.0"
instructor = ">=1.0.0"
aws-lambda-powertools = { version = ">=3.0.0", extras = ["tracer"] }
pydantic-settings = "^2.13.1"
fastapi = ">=0.115.0"
uvicorn = { version = ">=0.30.0", extras = ["standard"] }
mangum = ">=0.17.0"
yoyo-migrations = ">=9.0.0"
resend = ">=2.0.0"
email-validator = ">=2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
